Uses size_t as the result of hash for table.
Since absl::HashOf() returns size_t, which is usually a 64bit uint,
better not to receive the value as uint32_t.

PiperOrigin-RevId: 808890306

+++ src/composer/table.cc
@@ -574,7 +574,7 @@ std::shared_ptr<const Table> TableManager::GetTable(
     const mozc::commands::Request& request,
     const mozc::config::Config& config) {
   // calculate the hash depending on the request and the config
-  const uint32_t hash =
+  const size_t hash =
       absl::HashOf(request.special_romanji_table(), config.preedit_method(),
                    config.punctuation_method(), config.symbol_method());
